Modifying the test
