--- test/core/packageDependencyTests/HowickMaker/pkg.json
@@ -1,1 +1,0 @@
-{"license":"","file_hash":null,"name":"HowickMaker","version":"0.4.1","description":"Tools for programming steel stud roll-forming machines with Dynamo","group":"","keywords":["howick"],"dependencies":[],"contents":"","engine_version":"2.0.1.5065","engine":"dynamo","engine_metadata":"","site_url":"","repository_url":"","contains_binaries":true,"node_libraries":["HowickMaker, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"]}
